Fix missing Int64 conversion in VIPSSource seek method
- Convert the return value in seek method from gint64 to Int64
- Completes the Linux compatibility fix for type conversions

+++ Sources/VIPS/Core/VIPSSource.swift
@@ -246,7 +246,7 @@ public class VIPSSource: VIPSObject {
         if result == -1 {
             throw VIPSError()
         }
-        return result
+        return Int64(result)
     }
 
     /// Seeks to an absolute position in the source.
